draft(examples): Move current file to archive
Sorry for raw pull request, but I'm afraid to forget if I postpone sharing, and the script covers pretty basic scenario for a macro - moving the current active file to a specific folder. If it looks valid in general to you, I'll try to put more effort of adding Examples/Macro_MoveCurrentFileToArchive.md page. Will also be grateful for advice of how to screencapture a gif properly.