citationsManager imports keywords as tags
I am not a JS or TS developer, But I managed to add a few lines so the keywords of a citation can be imported as tags.
This might require an option to disable this feature if one does not need it.

Also in your guide in [here](https://bagerbach.com/blog/how-i-read-research-papers-with-obsidian-and-zotero#creating-notes-in-obsidian-from-papers-in-zotero) need to add the keywords to the template.

p.s. in the guide the tag for naming the new note should be {{VALUE:fileName}} while it is {{VALUE:title}} as causes issue with invalid characters.